Add constructor to subprocessors
It seems like there comes a time in the life of every module
written in a purely functional style, where an argument is passed
through a dozen functions before finally being used.  When
that point is reached in ruby, one solution is to use an instance
variable instead.  That's what I'm doing here.  The `rails` argument
becomes `@rails`.